Blood_supply/notification_service/requirements.txt
Django
djangorestframework
kafka-python
